% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_from_survfit.R
\name{dataset_from_survfit}
\alias{dataset_from_survfit}
\title{Recreate a dataset from a survfit object}
\usage{
dataset_from_survfit(fit)
}
\arguments{
\item{fit}{A survfit object from the 'survival' package.}
}
\value{
The data.frame with variables: time, status, and strata (if applicable).
}
\description{
Recreate a dataset from a survfit object
}
\examples{
fit = survival::survfit(survival::Surv(time, status) ~ sex, data = survival::lung)
dataset_from_survfit(fit)
}
