% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsurvplotlm.R
\name{ggsurvplotlm}
\alias{ggsurvplotlm}
\title{Plot and adjusted Kaplan-Meier curve accounting for landmark analysis}
\usage{
ggsurvplotlm(fit, landmark_time = 6, landmark_label = "Landmark Time", ...)
}
\arguments{
\item{fit}{A survfit object from the 'survival' package.}

\item{landmark_time}{The time at which to perform the landmark analysis}

\item{landmark_label}{The label to use for the landmark time in the plot}

\item{...}{Additional arguments passed to [survminer::ggsurvplot()].}
}
\value{
The modified ggsurvplot object.
}
\description{
Plot and adjusted Kaplan-Meier curve accounting for landmark analysis
}
