% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmark-package.R
\docType{package}
\name{landmarked-package}
\alias{landmarked}
\alias{landmarked-package}
\title{landmarked: Plot Adjusted Kaplan-Meier Estimates from a Landmark Time}
\description{
Plots contextual landmark Kaplan-Meier curves. This extension of the `survminer::ggsurvplot()` function allows the specification of a landmark time and an optional label. The period before the landmark is displayed as a pooled survival curve, while curves beyond the landmark are presented according to the groups defined in the supplied survfit object.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CStats/landmarked}
}

}
\author{
\strong{Maintainer}: Chris Brown \email{c.brown@sydney.edu.au} (\href{https://orcid.org/0000-0003-2506-4871}{ORCID})

}
\keyword{internal}
